Doc tweaks
This patch does the following:

- Fixes all warnings in the doc tree
- Adds a modd + devd conf file that renders HTML and treats doc warnings as
errors. This lets you send warning output to your desktop notifier, and gives
livereload similar to the livehtml entry in the Makefile.
- Some small tweaks

Sphinx currently has a bug that causes unknown options to generate a warning:

https://github.com/sphinx-doc/sphinx/issues/2229

This is fixed in stable, but not yet released. You can update Sphinx to the
latest stable like so:

pip install --upgrade https://github.com/sphinx-doc/sphinx/archive/stable.zip