Remove support for Python 3.5
There are a number of significant improvements in Python3.6 - especially in
asyncio - that makes ditching Python 3.5 compelling. The next Ubuntu LTS will
be released before the next version of mitmproxy, and will include Python 3.6
in base.

This patch removes support for testing under Python 3.5 and changes our
documentation. There are deeper changes in the type system and so forth that we
will make over time.