use empty sentinel to indicate end of configuration
Without this patch the server replay module would be initialized, and
get its flows loaded, with incomplete options set. More concretely, due
to the ctx.options.changed blinker mitmproxy() will invoke the
"configure" trigger upon initializing the default configuration (or
something along these lines).

Because the "server_replay" value is present in this first iteration it
would automatically load the flows. We change the flow loading for the
server replay functionality to be at "the end" of configuration
initialization by using an empty "updated" tuple sentinel.