Remove over-zealous assert on metadata types
The only real rule here is that we want the metadata to be serializable, and
the current constraint fails in the face of compound types. We don't want to
re-implement a type checker here, so just remove the assert for now.

Once the session mechanism is in core, we'll have better ways to deal with
this.

Fixes #3180