Fix Flow being part of raw_format_flow cache key
raw_format_flow is annotated with the @lru_cache decorator, which
creates a cache of the function's return value based on its
parameters. Thus, assuming that the first argument (f) contains all
the information needed to compute the urwid control for the flow, this
effectively caches results. To make most use of this, remove the
second parameter representing the real Flow object to improve
@lru_cache's efficiency.