add MASQUERADE target for iptables
A MASQUERADE line is necessary for this to work.  I spent quite a while trying to figure out what was missing from the howto to make it work.  It looks to be safe to not specify any target interfaces or ips, etc, the host running mitmproxy can still be used for local traffic fine.