Add support for IPv6-only environments
Previously, the proxy would attempt to bind to an IPv6 + IPv4 enabled socket.
On failure, it would try to bind to an IPv4 only socket. If that failed, the
proxy would fail to start. This update makes it so that the proxy also tries
to bind to an IPv6-only socket, which is necessary in environments where IPv4
is disabled.

In short, the proxy will try binding in the following order, only moving
to the next step when binding fails:
IPv6 + IPv4 -> IPv4 only -> IPv6 only -> proxy fails to start.