Limit WebSocketFlow messages array to a fixed size
If mitmproxy/mitmdump is used with long-running (or even not very long) socket connections, memory usage grows continuously while a socket connection remains active. One approach could be to limit this array to a fixed size, so that only the last N messages are accessible.

It seems most of the user scripts interact with messages as they come in (vs. accessing the complete websocket message history), so this would likely not break any existing functionality.