webaddons: Respect the user’s preferred web browser (if it’s graphical)
The open_browser function used a hard-coded preference list to make
sure we don’t open a terminal browser like lynx on a headless server,
as this would suspend the process.  But that hard-coded list also made
us open the wrong browser if, e.g., both Chrome and Firefox are
installed and the user prefers Firefox.  Instead, always get the
user’s preferred browser, then check that it’s a browser that will run
in the background.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>