make the CA cert live longer (#4611)
We need to restrict `CERT_EXPIRY` because modern browsers only accept a ~ <= 1year period for certs, but re-reading all prior commits and issues as well as the CA/Browser Forum Baseline Requirements on the topic, it looks like we don't actually need to restrict our CA lifetime by that much. Let's try out a longer validity and see if anyone complains.