Allow passing a hostname which includes authentication credentials
This commit allows passing a hostname with authentication credentials as an upstream proxy value (e.g. upstream:http://user:pass@default-upstream-proxy:8080/) something which thus far would result in a connection error. This allows the connection address (which includes the credentials) to be used later to and set proxy authentication headers (amongst other things).

In addition, I believe that since the credentials could now be passed as part of the address, different credentials will prevent connection reuse resulting in a new connection when different credentials are used for the same hostname/port combination. I believe this is desired behavior since a proxy server can route traffic differently depending on the provided credentials.