savehar: Fix timings being set to null
The HAR spec says that the `connect` and `ssl` timings are set to -1
when they are not applicable. This is the case for HTTP (not HTTPS)
requests and for failed CONNECT requests, which are dumped to the HAR
file too.

This allows dumps with these requests to be parsed correctly on
harviewer.