Fix Hugo deprecation error (#7372)
* Update Hugo installed in CI from `v0.92.1` to `v0.139.3`

The version in CI is close to 3 years old, which means it does not
encounter the same problems that newer installations of Hugo do.

* Fix Hugo deprecation error

I got the following error when building the documentation with Hugo
`v0.139.3`:

```
ERROR deprecated: data.GetJSON was deprecated in Hugo v0.123.0 and will be removed in Hugo 0.140.0. use resources.Get or resources.GetRemote with transform.Unmarshal.
```

I fixed the error by implementing the change the error recommended.