Add portfile support
Convenient when manipulating mitmproxy programmatically, as it's often
useful to specify port 0 (for all sorts of servers) but then the user
needs a way to retrieve the actual port being used.

mitmproxy logs the bound ports out by default, however because it logs
to the fully buffered stdout *and* can generate quite a bit of log,
this tends to be inconvenient. Writing out a file at a user-specified
location makes it much easier for them to poll, and using json even if
there is a minor race the user can just keep reading until they see a
valid JSON file.