feat(webnext): add dedicated hook and component for flow table keyboard navigation
This change also fixes an issue where input events (e.g. the user typing something) of components in the flow details pane would be ignored due to the global key event listener 'taking over'. For instance, typing the letter 'd' in the comment field would have deleted the currently active flow instead of actually inserting the letter in the target field. This issue is fixed by localizing the key event handler to a specific component (e.g. the flow table) instead of globally.